stats_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
pika==1.3.2
